requirements.txt
asgiref
astroid
autopep8
certifi
charset-normalizer
coreapi
coreschema
dill
dj-database-url
Django
django-filter
djangorestframework
gunicorn
idna
isort
itypes
Jinja2
lazy-object-proxy
Markdown
MarkupSafe
mccabe
mysqlclient
platformdirs
psycopg2-binary
pycodestyle
Pygments
pylint
python-dotenv
pytz
PyYAML
requests
sqlparse
toml
tomli
tomlkit
uritemplate
urllib3
whitenoise
wrapt
